BMS driver clean-up
