+++ cob_bms_driver/CMakeLists.txt
@@ -15,8 +15,8 @@ catkin_package()
 ## Build ##
 ###########
 include_directories(
-  ${catkin_INCLUDE_DIRS}
   include
+  ${catkin_INCLUDE_DIRS}
 )
 
 add_executable(bms_driver_node src/cob_bms_driver_node.cpp)
